+++ Frends.Oracle.ExecuteProcedure/Frends.Oracle.ExecuteProcedure.Tests/UnitTests.cs
@@ -7,9 +7,47 @@ using Oracle.ManagedDataAccess.Client;
 namespace Frends.Oracle.ExecuteProcedure.Tests;
 
 [TestFixture]
-class UnitTests : ExecuteProcedureTestBase
+class UnitTests
 {
+    private static Connection _con;
+    private static Options _options;
+
+    private static string schema = "test_user";
+    private static string _connectionString = $"Data Source = (DESCRIPTION = (ADDRESS = (PROTOCOL = TCP)(HOST = localhost)(PORT = 51521))(CONNECT_DATA = (SERVICE_NAME = XEPDB1))); User Id = {schema}; Password={schema};";
+    private static string _connectionStringSys = "Data Source = (DESCRIPTION = (ADDRESS = (PROTOCOL = TCP)(HOST = localhost)(PORT = 51521))(CONNECT_DATA = (SERVICE_NAME = XEPDB1))); User Id = sys; Password=mysecurepassword; DBA PRIVILEGE=SYSDBA";
     private readonly string _proc = "unitestproc";
+
+    [OneTimeSetUp]
+    public void OneTimeSetup()
+    {
+        _con = new Connection
+        {
+            ConnectionString = _connectionString,
+            TimeoutSeconds = 30
+        };
+
+        _options = new Options
+        {
+            ThrowErrorOnFailure = true,
+            BindParameterByName = true
+        };
+
+        Helpers.TestConnectionBeforeRunningTests();
+
+        using var con = new OracleConnection(_connectionStringSys);
+        con.Open();
+        Helpers.CreateTestUser(con);
+        con.Close();
+    }
+
+    [OneTimeTearDown]
+    public void OneTimeTearDown()
+    {
+        using var con = new OracleConnection(_connectionStringSys);
+        con.Open();
+        Helpers.DropTestUser(con);
+        con.Close();
+    }
 
     [SetUp]
     public void Setup()
