+++ Frends.Oracle.ExecuteProcedure/Frends.Oracle.ExecuteProcedure/Definitions/Input.cs
@@ -1,6 +1,5 @@
 ﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using System.Data;
 
 namespace Frends.Oracle.ExecuteProcedure.Definitions;
