+++ Frends.Oracle.ExecuteProcedure/Frends.Oracle.ExecuteProcedure/Definitions/OracleCommandReturnType.cs
@@ -1,8 +1,9 @@
-﻿#pragma warning disable 1591
-namespace Frends.Oracle.ExecuteProcedure.Definitions
-{
-    public enum OracleCommandReturnType 
-    { 
-        XmlString, XDocument, AffectedRows, JSONString, Parameters 
-    }
+﻿// self explanatory enum attributes. 
+#pragma warning disable 1591
+namespace Frends.Oracle.ExecuteProcedure.Definitions;
+
+public enum OracleCommandReturnType 
+{ 
+    XmlString, XDocument, AffectedRows, JSONString, Parameters 
 }
+
